data-scraper/llm_tools/requirements.txt
python-dotenv>=1.0
httpx>=0.27
tenacity>=8.2
